requirement.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dill==0.3.4
easyocr==1.4.2
fastapi==0.75.2
h11==0.13.0
idna==3.3
imageio==2.18.0
multiprocess==0.70.12.2
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pydantic==1.9.0
pyparsing==3.0.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
threaded==4.1.0
tifffile==2022.4.28
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
